+++ src/constants/passenger.ts
@@ -1,0 +1,2 @@
+export const MIN_VALUE = 0;
+export const MAX_VALUE = 3;
